Deleted confirmation added

+++ src/components/MyCampaign/MyCampaign.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../Auth/AuthProvider";
 import { Link } from "react-router-dom";
 import { MdDeleteForever } from "react-icons/md";
+import Swal from "sweetalert2";
 
 const MyCampaign = () => {
   const { user } = useContext(AuthContext);
@@ -15,13 +16,35 @@ const MyCampaign = () => {
       });
   }, [userEmail]);
   const handleDelete = (id) => {
-    fetch(`http://localhost:5000/campaign/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data)
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // Delete user in database
+        fetch(`http://localhost:5000/campaign/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            // console.log(data);
+            if (data.deletedCount) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your campaign has been deleted.",
+                icon: "success",
+              });
+              const remaningUsers = myCampaign.filter((user) => user._id !== id);
+              setMyCampaign(remaningUsers);
+            }
+          });
+      }
+    });
   };
   return (
     <div className="w-full overflow-x-auto mt-6">
